requirements.txt
streamlit
pandas
plotly
numpy
matplotlib
seaborn
scikit-learn
pillow